fix Global Descriptor Table :oD

+++ Memory/structGDT.h
@@ -5,9 +5,9 @@ struct gdtEntry
 {
    uint16_t limitLow;           
    uint16_t baseLow;
-   uint16_t baseMiddle;                 
-   uint8_t  privilege;              // Access flags, determine what ring this segment can be used in.
-   uint8_t  granularity;      
+   uint8_t baseMiddle;                 
+   uint8_t privilege;              // Access flags, determine what ring this segment can be used in.
+   uint8_t granularity;      
    uint8_t baseHigh;   
 } __attribute__((packed));
 typedef struct gdtEntry gdtEntry_t; 
@@ -17,8 +17,8 @@ typedef struct gdtEntry gdtEntry_t;
 */
 struct gdtPtr
 {
-	uint16_t lastAddress;
-	uint32_t firstAddress;
+	unsigned short lastAddress;
+	unsigned int firstAddress;
 } __attribute__((packed));
 typedef struct gdtPtr gdtPtr_t;
